(bg):avoiding repatition
