--- server/src/util/joiresponse.js
@@ -1,10 +1,0 @@
-import errorResponse from "./errorResponse";
-
-export const joiResponse = (req, res, schema, next) => {
-  const { error } = schema.validate(req.body);
-  if (error)
-    return next(
-      new errorResponse(error.details[0].message.replace(/\"/g, ""), 404)
-    );
-  next();
-};
